+++ DataFormats/interface/Jet.h
@@ -22,7 +22,7 @@ namespace phys {
   public:
     
     /// Constructor
-    Jet(const TLorentzVector& p = TLorentzVector(0.,0.,0.,0.), int q =0, int id = 0)
+    Jet(const TLorentzVector& p = TLorentzVector(0.,0.,0.,0.), float q =0, int id = 0)
       : Particle(p, q, id)
       , nConstituents_(-1)
       , nCharged_(-1)
