+++ DataFormats/interface/Lepton.h
@@ -23,7 +23,7 @@ namespace phys {
   public:
     
     /// Constructor
-    Lepton(const TLorentzVector& p = TLorentzVector(0.,0.,0.,0.), int q =0, int id = 0)
+    Lepton(const TLorentzVector& p = TLorentzVector(0.,0.,0.,0.), float q =0, int id = 0)
       : Particle(p,q, id)
       , dxy_(-9999.)               
       , dz_(-9999.)
